+++ src/api/Api.jsx
@@ -23,3 +23,11 @@ export function apiCallToGetEmployees(teamId) {
 export function apiCallToGetEmployeeMetric(employeeId) {
   return Mock.Employees[0];
 }
+
+export function apiCallToGetEmployeeOffice(officeId) {
+  return Mock.Office;
+}
+
+export function apiCallToGetEmployeeWorkstation(workstationId) {
+  return Mock.Workstation;
+}