+++ src/components/Banner/Banner.styles.jsx
@@ -2,7 +2,7 @@ import { styled } from "styled-components";
 import { appwhite } from '../../utils/colors.styles.jsx';
 
 export const $Banner = styled.header`
-  position: sticky;
+  position: relative;
   top: 0;
   width: 100%;
   height: 100px;