+++ src/components/Main/Metrics/MetricsEmployee/MetricsEmployee.styles.jsx
@@ -1,14 +1,15 @@
 import { styled } from "styled-components";
 
 export const $MetricsEmployee = styled.div`
+  padding: 0 20px;
   display: grid;
   grid-auto-flow: column dense;
-  grid-template-columns: 1fr 1fr 1fr;
+  grid-template-columns: 1fr 1fr 1fr 1fr;
   grid-template-rows: 1fr 1fr;
   gap: 20px 20px;
   grid-template-areas:
-    "Employee Workstation Team"
-    "Office Workstation Team";
+    "Employee Workstation Carbon"
+    "Office Workstation Carbon";
 
   & :first-child {
     grid-area: Employee;
@@ -23,7 +24,6 @@ export const $MetricsEmployee = styled.div`
   }
 
   & :nth-child(4) {
-    grid-area: Team;
+    grid-area: Carbon;
   }
-
 `;