+++ CHANGELOG.md
@@ -1,11 +1,11 @@
-0.16 (2024-05-??)
+0.16 (2024-05-18)
 ===============
 * [@BartoszDziurowicz](https://github.com/BartekDziurowicz/carbon-tracker-web/pull/6)
 * Simplify data from server, all objects are not needed
 * Change object info to carbon sumamry with new style and tooltip
 * Add constants to handle threshold value and carbon usage value
 * Create mocked api calls 
-* Develop and style all employee info components
+* Develop and style all employee info components include office info and workstation
 
 0.15 (2024-05-15)
 ===============